Add ro proof

+++ .profile
@@ -42,20 +42,23 @@ esac
 # Unless you edit my config you don't need this anymore so unset it
 unset File Setup Check
 
-# My GitHub SSH.
-if [ -z "$SSH_AGENT_PID" -a $UID != 0 ]; then # Only spawn SSH-AGENT if it never spawn before
-	if eval `ssh-agent -s`; then # If `eval` success then do ssh-add
-		case "$USER" in
-			ides3rt) # If me(ides3rt) login then add my github's ssh
-				ssh-add $HOME/.ssh/GitHub ;;
+# Don't do anyting if root is mount as ro
+if [ -n "$(findmnt / | awk '{print $4}' | grep -Eo '^rw,')" ]; then
+	# My GitHub SSH.
+	if [ -z "$SSH_AGENT_PID" -a $UID != 0 ]; then # Only spawn SSH-AGENT if it never spawn before
+		if eval `ssh-agent -s`; then # If `eval` success then do ssh-add
+			case "$USER" in
+				ides3rt) # If me login then add my GitHub's SSH
+					ssh-add $HOME/.ssh/GitHub ;;
 
-			*) ;;
-		esac
-		trap 'kill $SSH_AGENT_PID' EXIT
-fi	fi
+				*) ;;
+			esac
+			trap 'kill $SSH_AGENT_PID' EXIT
+	fi	fi
 
-# Automatic `startx` when in tty1. It's conflict with your favorite DM
-if [ -z "$DISPLAY" -a "$(tty)" = '/dev/tty1' -a $UID != 0 ]; then
-	# You don't need to specify /path/to/xinitrc if you use default location
-	exec startx $HOME/.config/X11/xinitrc # `exec` is unrequire, but it's a good thing
+	# Automatic `startx` when in tty1. It's conflict with your favorite DM
+	if [ -z "$DISPLAY" -a "$(tty)" = '/dev/tty1' -a $UID != 0 ]; then
+		# You don't need to specify /path/to/xinitrc if you use default location
+		exec startx $HOME/.config/X11/xinitrc # `exec` is unrequire, but it's a good thing
+	fi
 fi
